kernel source says tcp ports are 16 bits - sounds right

+++ main.c
@@ -135,8 +135,8 @@ int main(int argc, char **argv)
 
   // TODO:  is this the right size for this?  VSCode very angry at `struct ip`
   // kmo 9 dec 2023 20h42
-  uint32_t source_port = tcp_header->source;
-  uint32_t dest_port = tcp_header->dest;
+  uint16_t source_port = tcp_header->source;
+  uint16_t dest_port = tcp_header->dest;
   LOG_INFO("source port: %d", source_port);
   LOG_INFO("destination port: %d", dest_port);
